+++ Tests/DeserializerTests/TOMLInvalidationTests.swift
@@ -617,4 +617,10 @@ final class TOMLInvalidationTests: XCTestCase {
         let tomlData = try Data(contentsOf: tomlURL)
         XCTAssertThrowsError(try TOMLDeserializer.tomlTable(with: tomlData))
     }
+
+    func test_bare_cr() throws {
+        let tomlURL = URL(fileURLWithPath: "\(self.directory)\(kSeparator)bare-cr.toml", isDirectory: false)
+        let tomlData = try Data(contentsOf: tomlURL)
+        XCTAssertThrowsError(try TOMLDeserializer.tomlTable(with: tomlData))
+    }
 }